modified:   MetaTradeNode/BlockchainService.h 	modified:   MetaTradeNode/CryptoUtils.h 	modified:   MetaTradeNode/LevelDBLocalImpl.cpp 	modified:   MetaTradeNode/LevelDBLocalImpl.h 	modified:   MetaTradeNode/LocalService.h 	modified:   MetaTradeNode/MetaTradeBlockchainImpl.cpp 	modified:   MetaTradeNode/MetaTradeBlockchainImpl.h 	modified:   MetaTradeNode/MetaTradeBlockchainModels.h 	modified:   MetaTradeNode/MetaTradeNode.cpp 	modified:   MetaTradeNode/StompConst.cpp 	new file:   MetaTradeNode/dbviewer.cpp 	modified:   MetaTradeNode/example.cpp 	new file:   MetaTradeNode/lib/leveldb_debug.lib 	modified:   dlls/leveldb.dll 	new file:   dlls/leveldb_debug.dll

+++ MetaTradeNode/MetaTradeBlockchainModels.h
@@ -27,7 +27,7 @@ namespace metatradenode {
 		int proof;
 		std::vector<Trade> block_body;
 		void calMerkleHash();
-		double getBlockCommision();
+		long getBlockCommision();
 		std::string getHash();
 	};
 }
